requirements.txt
qrcode
requests
Pillow
svgwrite